requirements.txt
aiohttp
aiofiles
ffmpeg-python
gTTS
dnspython
future
googletrans==4.0.0-rc1
motor
pillow
psutil
pykeyboard
git+https://github.com/amanqs/pyrogram
python-arq
requests
search-engine-parser
speedtest-cli
uvloop
youtube_dl
bs4
python-dotenv
feedparser
pyromod==1.5
fuzzysearch
img2pdf
telegraph
pytube
pytz
python-dotenv
openai
psutil
httpx[http2]
cachetools
pyrate-limiter
SpeechRecognition
py-extract
py-trans
pymediainfo
wget
youtube-search-python
yt-dlp
regex
Telethon==1.*
TgCrypto